fix: update postgrex to 0.20.0
