+++ test/realtime/tenants/connect_test.exs
@@ -317,24 +317,30 @@ defmodule Realtime.Tenants.ConnectTest do
 
       # This creates a loop of errors that occupies all WAL senders and lets us test the error handling
       pids =
-        for i <- 0..10 do
+        for i <- 0..4 do
           replication_slot_opts =
             %PostgresReplication{
               connection_opts: opts,
               table: :all,
               output_plugin: "pgoutput",
-              output_plugin_options: [],
-              handler_module: TestHandler,
+              output_plugin_options: [proto_version: "1", publication_names: "test_#{i}_publication"],
+              handler_module: Replication.TestHandler,
               publication_name: "test_#{i}_publication",
               replication_slot_name: "test_#{i}_slot"
             }
 
-          {:ok, pid} = PostgresReplication.start_link(replication_slot_opts)
-          pid
+          spawn(fn ->
+            {:ok, pid} = PostgresReplication.start_link(replication_slot_opts)
+
+            receive do
+              :stop -> Process.exit(pid, :kill)
+            end
+          end)
         end
 
       on_exit(fn ->
-        Enum.each(pids, &Process.exit(&1, :normal))
+        Enum.each(pids, &send(&1, :stop))
+        Process.sleep(2000)
       end)
 
       log =
